+++ CHANGELOG.rst
@@ -9,8 +9,9 @@ Changelog
 * Update :meth:`pyvcf.VcfFrame.subset` method to take string as input in addition to list.
 * Fix a minor bug in :meth:`pymaf.MafFrame.plot_snvclss` method when certain SNV classes are entirely missing.
 * Add ``statsmodels`` package as dependency for performing statistical analysis.
-* Update :meth:`pymaf.MafFrame.plot_regplot` to calculate and print summary statistics as well.
-* Update :meth:`pyvcf.VcfFrame.plot_regplot` to calculate and print summary statistics as well.
+* Update :meth:`pymaf.MafFrame.plot_regplot` method to calculate and print summary statistics as well.
+* Update :meth:`pyvcf.VcfFrame.plot_regplot` method to calculate and print summary statistics as well.
+* Add :meth:`pyvcf.VcfFrame.miss2ref` method.
 
 0.17.0 (2021-07-08)
 -------------------